refactor: model 파일의 위치 변경
- src/js/에서 src/js/models로 수정
