+++ src/js/__tests__/lottoResult.test.js
@@ -1,4 +1,4 @@
-import lottoStatisticMachine from '../lottoStatisticMachine.js';
+import lottoStatisticMachine from '../models/lottoStatisticMachine.js';
 
 describe('구매한 로또 번호와 지난주 당첨 번호, 보너스 번호를 이용해서 당첨 결과를 확인할 수 있어야 한다.', () => {
   test('구매한 로또 중 당첨된 로또를 개수를 등수 별로 계산할 수 있어야 한다.', () => {
